pyproject.toml
[tool.poetry]
name = "pyimport"
version = "1.8.2"
description = "A CSV importer for MongoDB"
authors = ["Joe Drumgoole"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pymongo = "^4.7.3"
python-dateutil = "^2.9.0.post0"
toml = "^0.10.2"
requests = "^2.32.2"
configargparse = "^1.7"
coverage = "^7.5.1"
tqdm = "^4.66.4"
motor = "^3.4.0"
aiocsv = "^1.3.2"
aiofile = "^3.8.8"
aiohttp = "^3.9.5"
pytest-asyncio = "^0.23.7"
asyncstdlib = "^3.12.3"
mimesis = "^17.0.0"
aiofiles = "^23.2.1"
python-dotenv = {extras = ["cli"], version = "^1.0.1"}
pandas = "^2.2.2"
pytz = "^2024.1"
py-dateinfer = "^0.4.5"
sqlalchemy = "^2.0.34"
psycopg2-binary = "^2.9.9"
dnspython = "^2.6.1"
pytest = "^8.3.2"
tox = "^4.18.0"
psycopg = "^3.2.1"
sphinx = "^8.0.2"
myst-parser = "^4.0.0"
invoke = "^2.2.0"

[tool.poetry.scripts]
pyimport = "pyimport.pyimport_main:pyimport_main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
